requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
google-generativeai==0.3.2
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
